moving to async createHandlebars function to do dynamic import
