--- index.d.ts
@@ -1,4 +1,0 @@
-import * as Handlebars from 'handlebars';
-import { handlebarHelpers } from './handlebarHelpers';
-export = Handlebars;
-export { handlebarHelpers };
