--- index.js
@@ -1,8 +1,0 @@
-const HandlebarsLib = require('handlebars');
-const helpers = require('./helpers.js');
-
-const handlebars = HandlebarsLib.create();
-helpers({ handlebars: handlebars });
-
-module.exports = handlebars;
-module.exports.handlebarHelpers = helpers;
