+++ test/es6.mjs
@@ -1,8 +1,10 @@
-import handlebars from '../index.js';
+import {createHandlebars, helpers, handlebars} from '../dist/index.js';
 import {expect} from 'chai';
 
 describe('testing es6 examples', function() {
-    it('should get the handlebars from an es6 import', function() {
+    it('should get the handlebars from an es6 import', async function() {
+        const handlebars = await createHandlebars();
+
         // Use Fumanchu instead of Handlebars
         const source = 'Hello, {{name}}!';
         const template = handlebars.compile(source);
@@ -13,7 +15,9 @@ describe('testing es6 examples', function() {
         expect(result).to.equal('Hello, John!');
         
         //testing the export of handlebarHelpers
-        const {handlebarHelpers} = handlebars;
-        expect(handlebarHelpers).to.be.a('function');
+        expect(helpers).to.be.a('function');
+
+        //testing the export of handlebarHelpers
+        expect(handlebars).to.not.be.undefined;
     });
 });
