+++ test/index.js
@@ -1,10 +1,12 @@
 var assert = require('assert');
 const chai = require('chai');
 const expect = chai.expect;
-const Fumanchu = require('../index.js');
+const {createHandlebars} = require('../dist/index.js');
 
 describe('Fumanchu Template Test', function() {
-    it('should render name correctly', function() {
+    it('should render name correctly', async function() {
+        const Fumanchu = await createHandlebars();
+
         // Use Fumanchu instead of Handlebars
         const source = 'Hello, {{name}}!';
         const template = Fumanchu.compile(source);
@@ -15,13 +17,14 @@ describe('Fumanchu Template Test', function() {
         expect(result).to.equal('Hello, John!');
     });
 
-    it('should return the sum of two numbers.', function() {
+    it('should return the sum of two numbers.', async function() {
+        const Fumanchu = await createHandlebars();
         var fn = Fumanchu.compile('{{add value 5}}');
         assert.equal(fn({value: 5}), '10');
     });
 
     it('should have the helper class exported', function() {
-        var helpers = require('../index.js').handlebarHelpers;
+        var helpers = require('../dist/index.js').helpers;
         const HandlebarsLib = require('handlebars');
         const handlebars = HandlebarsLib.create();
         helpers({ handlebars: handlebars });
