+++ tsconfig.json
@@ -1,0 +1,22 @@
+{
+    "compilerOptions": {
+        "target": "ES6", // or your preferred target version
+        "module": "commonjs", // or your preferred module type
+        "allowJs": true, // Enable JavaScript files in the project
+        "checkJs": false, // Set to true if you want type-checking for JS files
+        "outDir": "./dist", // Specify output directory
+        "strict": true, // Enable strict type checking options
+        "esModuleInterop": true, // Required if you are using ES modules
+        "skipLibCheck": true, // Skipping type checks on declaration files
+        "declaration": true // Generate corresponding .d.ts file
+    },
+    "include": [
+        "helpers.js",
+        "src/**/*.js",
+        "index.ts"
+    ],
+    "exclude": [
+        "node_modules",
+        "test/**/*.js", // Exclude test files or other JS files you don't want to compile
+    ]
+}
